Implement JST/Handlebars as concat() filters.
This is a much better approach, no longer requires abusing multiple filter
functions to work in concert, fixes the caching issue of #188, and in general
should save us some headache.

As part of this, FilterTool.apply_func was changed to raise an exception if no
filters run. This allows us to differentiate between no filter running and a
filter not returning anything. We didn't end up needing this here in the end,
but it is useful nonetheless.